change default setting
