--- backend/src/config/config.ts
@@ -1,21 +1,0 @@
-// ルームの最大人数
-export const MAX_PLAYER = 5;
-
-// 初期に設置できる爆弾の数
-export const INITIAL_SET_BOMB_NUM = 1;
-
-// 初期の爆弾の破壊力
-export const INITIAL_BOMB_STRENGTH = 1;
-
-// プレイヤーの移動速度
-export const INITIAL_PLAYER_SPEED = 1;
-
-// プレイヤーの初期位置
-// TODO: サイズから計算する
-export const INITIAL_PLAYER_POSITION = [
-  { x: 96, y: 156 },
-  { x: 864, y: 156 },
-  { x: 96, y: 796 },
-  { x: 864, y: 796 },
-  { x: 480, y: 476 },
-];
