+++ backend/src/constants/constants.ts
@@ -1,4 +1,4 @@
-export const FPS = 120; // 1 秒間のフレーム数
+export const FPS = 60; // 1 秒間のフレーム数
 export const FRAME_RATE = 1000 / FPS; // 1 frame にかかる時間(ms)
 
 // Dockerfile の中と、デプロイ時にポートを指定しているので、ここの設定は開発時にしか利用されません。
@@ -31,3 +31,25 @@ export const GAME_STATE = {
 } as const;
 
 export type GAME_STATE_TYPE = typeof GAME_STATE[keyof typeof GAME_STATE];
+
+// ルームの最大人数
+export const MAX_PLAYER = 4;
+
+// 初期に設置できる爆弾の数
+export const INITIAL_SET_BOMB_NUM = 1;
+
+// 初期の爆弾の破壊力
+export const INITIAL_BOMB_STRENGTH = 1;
+
+// プレイヤーの初期移動速度
+export const INITIAL_PLAYER_SPEED = 4;
+
+// プレイヤーの初期位置
+// TODO: サイズから計算する
+export const INITIAL_PLAYER_POSITION = [
+  { x: 96, y: 156 },
+  { x: 864, y: 156 },
+  { x: 96, y: 796 },
+  { x: 864, y: 796 },
+  { x: 480, y: 476 },
+];
