+++ frontend/src/characters/Player.ts
@@ -1,11 +1,12 @@
 import Phaser from 'phaser';
 
+import * as Constants from '../../../backend/src/constants/constants';
 import IngameConfig from '../config/ingameConfig';
 import { handleCollide } from '../utils/handleCollide';
 
 export default class Player extends Phaser.Physics.Matter.Sprite {
-  public speed = 1;
-  public bombStrength = 1;
+  public speed: number;
+  public bombStrength: number;
 
   constructor(
     world: Phaser.Physics.Matter.World,
@@ -16,16 +17,20 @@ export default class Player extends Phaser.Physics.Matter.Sprite {
     options?: Phaser.Types.Physics.Matter.MatterBodyConfig
   ) {
     super(world, x, y, texture, frame, options);
+
+    this.speed = Constants.INITIAL_PLAYER_SPEED;
+    this.bombStrength = Constants.INITIAL_BOMB_STRENGTH;
+
     this.setScale(1, 1);
     this.setRectangle(IngameConfig.defaultTipSize, IngameConfig.defaultTipSize, {
-      chamfer: 10,
+      chamfer: 0,
       friction: 0,
       frictionStatic: 0,
       frictionAir: 0,
     });
     this.setOrigin(0.5, 0.5);
     this.setFixedRotation();
-    this.setSpeed(1);
+    this.setSpeed(this.speed);
     this.play('player_down', true); // 最初は下向いてる
 
     this.setOnCollide((data: Phaser.Types.Physics.Matter.MatterCollisionData) => {
